Unsealing the factory argument attribute.
I tried using this with a 'standard factories' static class, and needing to say:
  [CombinatorialFactoryArgument(0, Factories.Name, FactoryDeclaringType = typeof(Factories))]
for each argument is *very* cumbersome.

Unsealing this type allows client code to create overrides which would allow for nicer syntax, such as:
  [StandardFactoryArgument(0, Factories.Name)]
where the attribute just sets the declaring type.

I also refactored how the factory method is resolved to allow for more customization for factory arguments.