Reworked combinatorial argument attributes.
So I had a bit of an "oh wait, I'm being dumb" moment while writing up the last commit's message, and I remembered that attributes can be applied to parameters directly.

Re-worked how the combinatorial arguments are applied so that they are on the parameters themselves.  In doing so, I removed some extra unnecessary redundancy (eg: you no longer need to give the enum type to [CombinatorialEnum] uses.

Also re-worked the CombinatorialTestMethodAttribute accordingly.